setup test requirements and rename carriers -> providers
