+++ purpleserver/core/purpleserver/core/gateway.py
@@ -8,7 +8,7 @@ from rest_framework.exceptions import NotFound
 from purplship import package as api
 from purplship.core.utils import exec_async, to_dict
 
-from purpleserver.carriers import models
+from purpleserver.providers import models
 from purpleserver.core.exceptions import PurplShipApiException
 from purpleserver.core.datatypes import (
     CarrierSettings, ShipmentRequest, ShipmentResponse, RateRequest, Shipment,
