+++ purpleserver/core/purpleserver/providers/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-import purpleserver.carriers.models as models
+import purpleserver.providers.models as models
 
 for name, model in models.MODELS.items():
     admin.site.register(model)
