+++ purpleserver/setup.py
@@ -28,7 +28,7 @@ manager_requirements = [
 
 setup(
       name='purplship-server',
-      version='2020.7.1',
+      version='2020.8.0',
       description='Multi-carrier shipping API',
       long_description=long_description,
       long_description_content_type="text/markdown",
