+++ requirements.test.txt
@@ -1,0 +1,8 @@
+-f https://git.io/purplship
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship.canadapost&subdirectory=extensions/canadapost
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship.dhl&subdirectory=extensions/dhl
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship.fedex&subdirectory=extensions/fedex
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship.package&subdirectory=extensions/package
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship.purolator&subdirectory=extensions/purolator
+-e git+https://github.com/PurplShip/purplship.git@release-2020.8.0#egg=purplship.ups&subdirectory=extensions/ups
